organizing the content a bit more
