+++ sample/web/private.go
@@ -1,0 +1,15 @@
+package web
+
+import "net/http"
+
+func private(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "text/html")
+	content, err := templates.ReadFile("templates/private.html")
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+
+		return
+	}
+
+	w.Write(content)
+}
