+++ sample/web/public.go
@@ -1,0 +1,15 @@
+package web
+
+import "net/http"
+
+func public(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "text/html")
+	content, err := templates.ReadFile("templates/public.html")
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+
+		return
+	}
+
+	w.Write(content)
+}
